Исправил AbstractFileStorage. Рекурсия в одном файле. Вывод pathOutput() только имя файла. Сгенерировал equals() и hashCode(), где их не было.

+++ src/com/urise/webapp/storage/AbstractFileStorage.java
@@ -3,7 +3,9 @@ package com.urise.webapp.storage;
 import com.urise.webapp.exception.StorageException;
 import com.urise.webapp.model.Resume;
 
-import java.io.*;
+import java.io.File;
+import java.io.IOException;
+import java.util.ArrayList;
 import java.util.List;
 import java.util.Objects;
 
@@ -35,35 +37,45 @@ public abstract class AbstractFileStorage extends AbstractStorage<File> {
     protected void saveResume(Resume resume, File file) {
         try {
             file.createNewFile();
-            writeTo(resume, file);
         } catch (IOException e) {
-            throw new StorageException("IO error", file.getName(), e);
+            throw new StorageException("Couldn't create file " + file.getAbsolutePath(), file.getName(), e);
         }
+        updateResume(resume, file);
     }
 
     protected abstract void writeTo(Resume resume, File file) throws IOException;
 
+    protected abstract Resume readResume(File file) throws IOException;
+    //Вариант реализации readResume:
+//    protected abstract Resume readResume(File file) throws IOException {
+//        try {
+//            // Coздaem oбъekт FileInputStream для дocтyпa k фaйлy
+//            FileInputStream fin = new FileInputStream(directory.getAbsolutePath());
+//            // Coздaem oбъekт ObjectInputStream для чтeния cepиaлизoвaннoгo oбъekтa
+//            ObjectInputStream ois = new ObjectInputStream(fin);
+//            for (File resumeFile : Objects.requireNonNull(directory.listFiles())) {
+//                if (resumeFile.equals(file)) {
+//                    // Вoccтaнaвливaem cocтoяниe oбъekтa из фaйлa
+//                    return (Resume) ois.readObject();
+//                }
+//            }
+//        } catch (FileNotFoundException e) {
+//            throw new RuntimeException("File Not Found", e);
+//        } catch (IOException e) {
+//            throw new RuntimeException("IOException", e);
+//        } catch (ClassNotFoundException e) {
+//            throw new RuntimeException("Class Not Found", e);
+//        }
+//        return null;
+//    }
+
     @Override
     protected Resume getResume(File file) {
         try {
-            // Coздaem oбъekт FileInputStream для дocтyпa k фaйлy
-            FileInputStream fin = new FileInputStream(directory.getAbsolutePath());
-            // Coздaem oбъekт ObjectInputStream для чтeния cepиaлизoвaннoгo oбъekтa
-            ObjectInputStream ois = new ObjectInputStream(fin);
-            for (File resumeFile : Objects.requireNonNull(directory.listFiles())) {
-                if (resumeFile.equals(file)) {
-                    // Вoccтaнaвливaem cocтoяниe oбъekтa из фaйлa
-                    return (Resume) ois.readObject();
-                }
-            }
-        } catch (FileNotFoundException e) {
-            throw new RuntimeException("File Not Found", e);
+            return readResume(file);
         } catch (IOException e) {
-            throw new RuntimeException("IOException", e);
-        } catch (ClassNotFoundException e) {
-            throw new RuntimeException("Class Not Found", e);
+            throw new StorageException("File read error", file.getName(), e);
         }
-        return null;
     }
 
     @Override
@@ -71,37 +83,37 @@ public abstract class AbstractFileStorage extends AbstractStorage<File> {
         try {
             writeTo(resume, file);
         } catch (IOException e) {
-            throw new StorageException("IO error", file.getName(), e);
+            throw new StorageException("File write error", resume.getUuid(), e);
         }
     }
 
     @Override
     protected void deleteResume(File file) {
+        if (!file.delete()) {
+            throw new StorageException("File delete error", file.getName());
+        }
         file.delete();
     }
 
     @Override
     protected void clearStorage() {
         for (File file : Objects.requireNonNull(directory.listFiles())) {
-            file.delete();
+            deleteResume(file);
         }
     }
 
     @Override
     protected List<Resume> toList() {
-        try {
-            // Coздaem oбъekт FileInputStream для дocтyпa k фaйлy
-            FileInputStream fin = new FileInputStream(directory.getAbsolutePath());
-            // Coздaem oбъekт ObjectInputStream для чтeния cepиaлизoвaннoгo oбъekтa
-            ObjectInputStream ois = new ObjectInputStream(fin);
-            return (List<Resume>) ois.readObject();
-        } catch (FileNotFoundException e) {
-            throw new RuntimeException("File Not Found", e);
-        } catch (IOException e) {
-            throw new RuntimeException("IOException", e);
-        } catch (ClassNotFoundException e) {
-            throw new RuntimeException("Class Not Found", e);
+        File[] files = directory.listFiles();
+        if (files == null) {
+            throw new StorageException("Directory read error", null);
         }
+
+        List<Resume> list = new ArrayList<>(files.length);
+        for (File file : files) {
+            list.add(getResume(file));
+        }
+        return list;
     }
 
     @Override
